detach: new plugin for detached head
this will hopefully speed up git push. till now we did not know where we
were after a push, so we needed to check every file if it has to be
updated. now we store the old sha1 in the pre-receive, so that we have
to update only the modified files. much faster.